WIP (move the options to a debug sub-menu)
