+++ src/components/options/useCommentOptions.ts
@@ -19,13 +19,13 @@ import { useCachedLinks } from "components/hooks/useCachedLinks";
 import { useEditorsState } from "components/core/editors/EditorsContext";
 import { useThreadContext } from "components/thread/ThreadContext";
 import { useThreadEditors } from "components/core/editors/withEditors";
+import { useDebugOptions } from "./comment/useDebugOptions";
 
 export enum CommentOptions {
   GO_TO_COMMENT = "GO_TO_COMMENT",
-  COPY_CONTENT_DATA = "COPY_CONTENT_DATA",
-  COPY_CHAIN_CONTENT_DATA = "COPY_CHAIN_CONTENT_DATA",
   COPY_COMMENT_LINK = "COPY_COMMENT_LINK",
   REPLY_TO_COMMENT = "REPLY_TO_COMMENT",
+  DEBUG = "DEBUG",
 }
 
 /**
@@ -73,33 +73,6 @@ export const useCommentOptions = ({
       options
         .map((option) => {
           switch (option) {
-            case CommentOptions.COPY_CONTENT_DATA: {
-              return {
-                icon: faCopy,
-                name: "Copy content data",
-                link: {
-                  onClick: () => {
-                    copyText(comment.content);
-                    toast.success("Data copied!");
-                  },
-                },
-              };
-            }
-            case CommentOptions.COPY_CHAIN_CONTENT_DATA: {
-              return {
-                icon: faCopy,
-                name: "Copy comment chain content data",
-                link: {
-                  onClick: () => {
-                    const chainContent = JSON.stringify(
-                      commentChain.map((comment) => JSON.parse(comment.content)),
-                    );
-                    copyText(chainContent);
-                    toast.success("Data copied!");
-                  },
-                },
-              };
-            }
             case CommentOptions.COPY_COMMENT_LINK: {
               return {
                 icon: faLink,
@@ -136,6 +109,9 @@ export const useCommentOptions = ({
                   label: "Add a new comment",
                 },
               };
+            }
+            case CommentOptions.DEBUG: {
+              return useDebugOptions({ comment, commentChain });
             }
           }
         })
